move fixtures to conftest.py file
